Replace unused file and fix tests
