+++ apps/src/lib/kits/maker/ui/EligibilityChecklist.story.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import EligibilityChecklist from './EligibilityChecklist';
-import EligibilityChecklist2019 from './EligibilityChecklist2019';
 import {Status} from '@cdo/apps/lib/ui/ValidationStep';
 
 const defaultProps = {
@@ -20,7 +19,7 @@ export default storybook => {
         description: 'New format for 2019',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
             />
           </div>
@@ -31,7 +30,7 @@ export default storybook => {
         description: 'When your school does not qualify',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
               schoolId="1234"
               schoolName="Code.org Junior Academy"
@@ -46,7 +45,7 @@ export default storybook => {
         description: 'When your school does qualify',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
               schoolId="1234"
               schoolName="Code.org Junior Academy"
@@ -63,7 +62,7 @@ export default storybook => {
         description: 'When your school does qualify',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
               schoolId="1234"
               schoolName="Code.org Junior Academy"
@@ -78,7 +77,7 @@ export default storybook => {
         description: 'When you are not planning to teach this or next year',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
               schoolId="1234"
               schoolName="Code.org Junior Academy"
@@ -94,7 +93,7 @@ export default storybook => {
         description: 'When you are planning to teach this or next year',
         story: () => (
           <div style={{margin: '2em'}}>
-            <EligibilityChecklist2019
+            <EligibilityChecklist
               {...defaultProps}
               schoolId="1234"
               schoolName="Code.org Junior Academy"
@@ -103,103 +102,6 @@ export default storybook => {
               unit6Intention="yes1819"
             />
           </div>
-        )
-      },
-      {
-        name: 'Failed Checklist',
-        description: 'EligbilityChecklist where one of first list items failed',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            statusStudentCount={Status.FAILED}
-          />
-        )
-      },
-      {
-        name: 'Check Year Checklist',
-        description: 'First two items succeeded, third needs to be verified',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-          />
-        )
-      },
-      {
-        name: 'Ineligible year submitted',
-        description: 'User had submitted an ineligible response for unit 6 intentions',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            unit6Intention="no"
-          />
-        )
-      },
-      {
-        name: 'Eligible year submitted, user does not have a school',
-        description: 'User had submitted an eligible response for unit 6 intentions',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            unit6Intention="yes1718"
-          />
-        )
-      },
-
-      // Ideally we would have a story here for when the user has a school, but has
-      // not confirmed it for this application, however we dont end up with any schools
-      // in our dropdown in storybook
-
-      {
-        name: 'User has confirmed school',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            unit6Intention="yes1718"
-            schoolId="1234"
-            schoolName="Code.org Junior Academy"
-            hasConfirmedSchool={true}
-          />
-        )
-      },
-
-      {
-        name: 'Admin override before user filled anything out',
-        description: 'All bubbles should be green. We should have a get code button',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            unit6Intention=""
-            schoolId=""
-            schoolName=""
-            getsFullDiscount={true}
-            initialDiscountCode={null}
-            adminSetStatus={true}
-          />
-        )
-      },
-
-      {
-        name: 'Admin override after user answered unit6 intention',
-        description: 'Should still have green bubbles because of override',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            unit6Intention="no"
-            schoolId=""
-            schoolName=""
-            getsFullDiscount={false}
-            initialDiscountCode={null}
-            adminSetStatus={true}
-          />
-        )
-      },
-      {
-        name: 'Not currently distributing discount codes',
-        story: () => (
-          <EligibilityChecklist
-            {...defaultProps}
-            currentlyDistributingDiscountCodes={false}
-          />
         )
       },
     ]);
