+++ apps/src/sites/studio/pages/maker/discountcode.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import EligibilityChecklist2019 from '@cdo/apps/lib/kits/maker/ui/EligibilityChecklist2019';
+import EligibilityChecklist from '@cdo/apps/lib/kits/maker/ui/EligibilityChecklist';
 import DiscountAdminOverride from '@cdo/apps/lib/kits/maker/ui/DiscountAdminOverride';
 import { Status } from '@cdo/apps/lib/ui/ValidationStep';
 import getScriptData from '@cdo/apps/util/getScriptData';
@@ -16,7 +16,7 @@ $(document).ready(() => {
         <DiscountAdminOverride/>
       }
       {!is_admin &&
-        <EligibilityChecklist2019
+        <EligibilityChecklist
           statusPD={application.is_pd_eligible ? Status.SUCCEEDED : Status.FAILED}
           statusStudentCount={application.is_progress_eligible ? Status.SUCCEEDED : Status.FAILED}
           unit6Intention={application.unit_6_intention}
