+++ apps/test/unit/lib/kits/maker/ui/EligibilityChecklistTest.js
@@ -3,7 +3,6 @@ import {shallow} from 'enzyme';
 import {assert} from '../../../../../util/configuredChai';
 import EligibilityChecklist from '@cdo/apps/lib/kits/maker/ui/EligibilityChecklist';
 import {Status} from '@cdo/apps/lib/ui/ValidationStep';
-import i18n from "@cdo/locale";
 
 
 describe('EligibilityChecklist', () => {
@@ -24,47 +23,48 @@ describe('EligibilityChecklist', () => {
     assert(wrapper.is('div'));
   });
 
-  it('does not render DiscountCodeSchoolChoice until we answer unit6 question', () => {
+  it('does not render Unit6ValidationStep until we answer school choice question', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
       />
     );
-    assert.equal(wrapper.find('DiscountCodeSchoolChoice').length, 0);
-    wrapper.instance().handleUnit6Submitted(true);
-    assert.equal(wrapper.find('DiscountCodeSchoolChoice').length, 1);
+    assert.equal(wrapper.find('Unit6ValidationStep').length, 0);
+    wrapper.instance().handleSchoolConfirmed({schoolId: '1', fullDiscount: true});
+    assert.equal(wrapper.find('Unit6ValidationStep').length, 1);
   });
 
-  it('renders DiscountCodeSchoolChoice if we previously answered unit6 question', () => {
+  it('renders Unit6ValidationStep if we previously answered school choice question', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        hasConfirmedSchool
+        schoolId="1"
+        getsFullDiscount={true}
       />
     );
-    assert.equal(wrapper.find('DiscountCodeSchoolChoice').length, 1);
+    assert.equal(wrapper.find('Unit6ValidationStep').length, 1);
   });
 
-  it('does not render get code button until we confirm school', () => {
+  it('does not render get code button until we confirm school and unit 6 intention', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
         schoolId="12345"
         schoolName="Code.org Junior Academy"
-        hasConfirmedSchool={false}
+        hasConfirmedSchool={true}
       />
     );
     assert.equal(wrapper.find('Button').length, 0);
-    wrapper.instance().handleSchoolConfirmed(true);
+    wrapper.instance().handleUnit6Submitted(true);
     assert.equal(wrapper.find('Button').length, 1);
   });
 
-  it('renders get code button if we previously confirmed school', () => {
+  it('renders get code button if we previously confirmed school and unit 6 intention', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        unit6Intention="yes1819"
         schoolId="12345"
         schoolName="Code.org Junior Academy"
         hasConfirmedSchool={true}
@@ -78,7 +78,7 @@ describe('EligibilityChecklist', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        unit6Intention="yes1819"
         schoolId="12345"
         schoolName="Code.org Junior Academy"
         hasConfirmedSchool={true}
@@ -95,7 +95,7 @@ describe('EligibilityChecklist', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        unit6Intention="yes1819"
         schoolId="12345"
         schoolName="Code.org Junior Academy"
         hasConfirmedSchool={true}
@@ -111,7 +111,7 @@ describe('EligibilityChecklist', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        unit6Intention="yes1819"
         initialDiscountCode="MYCODE"
         initialExpiration="2018-12-31T00:00:00.000Z"
         getsFullDiscount={false}
@@ -120,16 +120,16 @@ describe('EligibilityChecklist', () => {
     assert(wrapper.is('DiscountCodeInstructions'));
   });
 
-  it('does not show Unit6ValidationStep radio buttons when it has admin override', () => {
+  it('does not show Unit6ValidationStep when it has admin override', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
-        unit6Intention="yes1718"
+        unit6Intention="yes1819"
         adminSetStatus={true}
         getsFullDiscount={true}
       />
     );
-    assert.equal(wrapper.find('Unit6ValidationStep').props().showRadioButtons, false);
+    assert.equal(wrapper.find('Unit6ValidationStep').length, 0);
   });
 
   it('claims unit6 success when it has admin override', () => {
@@ -141,10 +141,10 @@ describe('EligibilityChecklist', () => {
         getsFullDiscount={true}
       />
     );
-    assert.equal(wrapper.find('Unit6ValidationStep').props().stepStatus, Status.SUCCEEDED);
+    assert.equal(wrapper.state().statusYear, Status.SUCCEEDED);
   });
 
-  it('shows button to get code when admin override with full discount', () => {
+  it('shows button to get code when admin override', () => {
     const wrapper = shallow(
       <EligibilityChecklist
         {...defaultProps}
@@ -153,58 +153,6 @@ describe('EligibilityChecklist', () => {
         getsFullDiscount={true}
       />
     );
-    assert.equal(wrapper.find('Button').props().text, 'Get Code for $0 kit');
-  });
-
-  it('shows button to get code when admin override with partial discount', () => {
-    const wrapper = shallow(
-      <EligibilityChecklist
-        {...defaultProps}
-        unit6Intention="no"
-        adminSetStatus={true}
-        getsFullDiscount={false}
-      />
-    );
-    assert.equal(wrapper.find('Button').props().text, 'Get Code for $97.50 kit');
-  });
-
-  it('shows message about subsidized kits when admin override', () => {
-    const wrapper = shallow(
-      <EligibilityChecklist
-        {...defaultProps}
-        unit6Intention="no"
-        adminSetStatus={true}
-        getsFullDiscount={true}
-      />
-    );
-    assert.equal(wrapper.find('div').last().text(), i18n.eligibilityReqYearFail());
-  });
-
-  it('shows discount message when partial discount', () => {
-    const wrapper = shallow(
-      <EligibilityChecklist
-        {...defaultProps}
-        unit6Intention="yes1718"
-        schoolId="12345"
-        schoolName="Code.org Junior Academy"
-        hasConfirmedSchool={true}
-        getsFullDiscount={false}
-      />
-    );
-    assert(wrapper.text().includes("According to our data, your school has fewer than 50% of students"));
-  });
-
-  it('does not show discount message when full discount', () => {
-    const wrapper = shallow(
-      <EligibilityChecklist
-        {...defaultProps}
-        unit6Intention="yes1718"
-        schoolId="12345"
-        schoolName="Code.org Junior Academy"
-        hasConfirmedSchool={true}
-        getsFullDiscount={true}
-      />
-    );
-    assert(!wrapper.text().includes("According to our data, your school has fewer than 50% of students"));
+    assert.equal(wrapper.find('Button').props().text, 'Get Code');
   });
 });
